[Feat] 히어로섹션 인트로 애니메이션 통합 및 최적화
- 기존 인트로 컴포넌트를 통합하여 코드 간소화
- 마운트 상태에 따라 블랙 스크린 표시 추가
