+++ src/app/(route)/home/_components/HeroTitleSection/index.tsx
@@ -15,8 +15,7 @@ import Image from "next/image";
 import "swiper/css";
 import "swiper/css/effect-fade";
 import "swiper/css/autoplay";
-import Intro from "./Intro";
-import { getCookie } from "cookies-next";
+import { setCookie, getCookie } from "cookies-next";
 import useHeroImages from "../../_hook/useHeroImages";
 
 // 애니메이션 상수
@@ -60,12 +59,55 @@ const ANIMATION_CONSTANTS = {
 export default function HeroTitleSection() {
   const ref = useRef(null);
   const [isMounted, setIsMounted] = useState(false);
+  const [isLoaded, setIsLoaded] = useState(false);
+  const [showIntro, setShowIntro] = useState(true);
+  const [welcomeStep, setWelcomeStep] = useState(1);
+  const [isFadingIntro, setIsFadingIntro] = useState(false);
+
   const { heroImages, isLoading } = useHeroImages();
 
   // 클라이언트 측에서만 실행되도록 마운트 상태 설정
   useEffect(() => {
     setIsMounted(true);
+
+    // 쿠키 확인해서 인트로를 보여줄지 결정
+    const loadedCookie = getCookie("isLoaded");
+    if (loadedCookie === "true") {
+      setIsLoaded(true);
+      setShowIntro(false);
+    }
   }, []);
+
+  // 인트로 애니메이션 효과 처리
+  useEffect(() => {
+    if (!isMounted || isLoaded) return;
+
+    // 환영 메시지 단계별 표시
+    const step1Timer = setTimeout(() => {
+      setWelcomeStep(2);
+    }, 1000);
+
+    const fadeTimer = setTimeout(() => {
+      setIsFadingIntro(true);
+    }, 3300);
+
+    const completeTimer = setTimeout(() => {
+      // 인트로 애니메이션 완료 후 메인 콘텐츠 표시
+      setIsLoaded(true);
+      setCookie("isLoaded", "true", { maxAge: 60 * 10 }); // 10분 동안 유효
+
+      // 인트로가 완전히 페이드 아웃된 후 인트로를 안 보이게 처리
+      setTimeout(() => {
+        setShowIntro(false);
+      }, 700); // 페이드 아웃 애니메이션 시간에 맞춤
+    }, 4000);
+
+    return () => {
+      clearTimeout(step1Timer);
+      clearTimeout(fadeTimer);
+      clearTimeout(completeTimer);
+    };
+  }, [isMounted, isLoaded]);
 
   const { scrollYProgress } = useScroll({
     target: ref,
@@ -75,9 +117,8 @@ export default function HeroTitleSection() {
   const y = useTransform(scrollYProgress, [0, 1], ["0%", "40%"]);
   const opacity = useTransform(scrollYProgress, [0, 0.5], [1, 0]);
 
-  if (!isMounted) return null;
-
-  const isLoaded = getCookie("isLoaded");
+  // 마운트되지 않은 경우 블랙 스크린 표시
+  if (!isMounted) return <div className="fixed inset-0 bg-black z-50"></div>;
 
   // isLoaded 상태에 따라 애니메이션 설정 선택
   const sectionAnimation = isLoaded
@@ -92,12 +133,87 @@ export default function HeroTitleSection() {
 
   return (
     <>
-      {!isLoaded && <Intro />}
+      {/* 인트로 애니메이션 */}
+      {showIntro && (
+        <motion.div
+          className="fixed inset-0 overflow-hidden bg-black flex items-center justify-center z-50"
+          initial={{ opacity: 1 }}
+          animate={{ opacity: isFadingIntro ? 0 : 1 }}
+          transition={{ duration: 0.7, ease: "easeInOut" }}
+          suppressHydrationWarning
+        >
+          {/* 로고 애니메이션 */}
+          <motion.div
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            className="absolute inset-0 z-0"
+          >
+            <motion.div
+              className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 opacity-10"
+              animate={{
+                scale: [1, 1.2, 1],
+                rotate: [0, 5, 0, -5, 0],
+              }}
+              transition={{ duration: 10, repeat: Infinity }}
+            >
+              <svg
+                width="600"
+                height="600"
+                viewBox="0 0 200 200"
+                className="text-red-600/20"
+                suppressHydrationWarning
+              >
+                <path
+                  fill="currentColor"
+                  d="M100 0 A100 100 0 0 1 100 200 A100 100 0 0 1 100 0z"
+                />
+              </svg>
+            </motion.div>
+          </motion.div>
+
+          {/* 웰컴 메시지 */}
+          <AnimatePresence mode="wait">
+            {welcomeStep === 1 && (
+              <motion.div
+                key="welcome-1"
+                className="text-center z-10"
+                initial={{ opacity: 0, y: 20 }}
+                animate={{ opacity: 1, y: 0 }}
+                exit={{ opacity: 0, y: -20 }}
+                transition={{ duration: 0.6 }}
+                suppressHydrationWarning
+              >
+                <h2 className="text-5xl md:text-6xl font-bold text-white tracking-tight drop-shadow-[0_0_15px_rgba(0,0,255,0.8)] md:drop-shadow-[0_0_35px_rgba(0,0,255,0.8)]">
+                  오늘은 캠퍼스로!
+                </h2>
+              </motion.div>
+            )}
+
+            {welcomeStep === 2 && (
+              <motion.div
+                key="welcome-2"
+                className="text-center z-10"
+                initial={{ opacity: 0, y: 20 }}
+                animate={{ opacity: 1, y: 0 }}
+                exit={{ opacity: 0, y: -20 }}
+                transition={{ duration: 0.6 }}
+                suppressHydrationWarning
+              >
+                <h2 className="text-5xl md:text-6xl font-bold text-white tracking-tight drop-shadow-[0_0_15px_rgba(0,0,255,0.8)] md:drop-shadow-[0_0_35px_rgba(0,0,255,0.8)]">
+                  내일은 열방으로!
+                </h2>
+              </motion.div>
+            )}
+          </AnimatePresence>
+        </motion.div>
+      )}
+
+      {/* 메인 컨텐츠 */}
       <AnimatePresence>
         <motion.section
           ref={ref}
           initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
+          animate={{ opacity: isLoaded ? 1 : 0 }}
           transition={sectionAnimation}
           className="relative h-screen overflow-hidden flex items-center justify-center"
           suppressHydrationWarning
